Use new sync locking technique
This requires a hardware change: the input csync signal must be
connected to an analog input pin.

The input and output vsync pulses are phase locked by constantly
adjusting the size of the vertical front porch.  The output is locked
to trail the input so that it never "catches up" with the capture
buffer being filled by the deinterlacer, which would result in a
screen tear.