Properly apply Bypass video mode from custom preset
Both the newly added call to setOutModeHdBypass(), and the existing
calls in applyPresets(), do not initialize adco->r/g/b_gain from
GBS::ADC_R/G/BGCTRL. This may or may not be worth changing, as I've
noticed odd behavior when initializing gain, both in resolutions and
Bypass, fixed and custom resolutions, and auto gain on and off. I'm
choosing to copy the existing code for the time being.