Reduce latency for frequency-based frame sync
Frequency-based framesync is more precise and doesn't drift back and
forth relative to the target phase, so we can safely run it at a lower
latency without risking tearing or falling a frame behind.

This should be 100% safe in regular operation in a fixed video mode,
and *mostly* safe when the game switches between 240p and 480i
(safe unless externalClockGenSyncInOutRate is called too late after a
mode switch, or fails).

We allocate 0.125 frames of latency in steady state. This should be more
than the (up to 0.071+ frames of) buffer we lose when input slows down
from 262.5 to 263 lines, before we respond and change our output rate
to match.