Reduce 480p output pixel clock to 64.8 MHz
Previously every scanline's active period was shorter than expected, so
images would show horizontally compressed and not fill the full width
of displays. If you tried to work around this with horizontal
expansion, the image would appear corrupted at the right side.

This makes every scanline take longer to show, so we can reduce the
htotal period of each scanline, and scale the image *down* to avoid
being too wide (which does *not* corrupt the image).