Fix chroma scanlines
I have verified that the UV deinterlacer registers written by
disableScanlines() match the default values before enableScanlines() is
called (in line-doubling mode).


This change (mostly) fixes the bug where blue areas of the image barely
have visible scanline gaps, because the scanline darkening is only
applied to the luma but not chroma channels of the image.